refs #1753 added functionality to set local variable values with API calls
