+++ Makefile.am
@@ -390,6 +390,8 @@ noinst_HEADERS = \
 	include/qore/intern/qore_number_private.h \
 	include/qore/intern/qore_date_private.h \
 	include/qore/intern/sql_statement_private.h \
+	include/qore/intern/ThreadLocalVariableData.h \
+	include/qore/intern/ThreadClosureVariableStack.h \
 	include/qore/intern/qore_program_private.h \
 	include/qore/intern/qore_socket_private.h \
 	include/qore/intern/QoreDir.h \
