+++ include/qore/intern/qore_program_private.h
@@ -90,311 +90,8 @@ public:
 
 typedef QoreThreadLocalStorage<QoreHashNode> qpgm_thread_local_storage_t;
 
-class ThreadLocalVariableData : public ThreadLocalData<LocalVarValue> {
-public:
-   // marks all variables as finalized on the stack
-   DLLLOCAL void finalize(arg_vec_t*& cl) {
-      ThreadLocalVariableData::iterator i(curr);
-      while (i.next()) {
-         AbstractQoreNode* n = i.get().finalize();
-         if (n && n->isReferenceCounted()) {
-            if (!cl)
-               cl = new arg_vec_t;
-            cl->push_back(n);
-         }
-      }
-   }
-
-   // deletes everything on the stack
-   DLLLOCAL void del(ExceptionSink* xsink) {
-      // then we uninstantiate
-      while (curr->prev || curr->pos)
-         uninstantiate(xsink);
-   }
-
-   DLLLOCAL LocalVarValue* instantiate() {
-      if (curr->pos == QORE_THREAD_STACK_BLOCK) {
-	 if (curr->next)
-	    curr = curr->next;
-	 else {
-	    curr->next = new Block(curr);
-	    //printf("this: %p: add curr: %p, curr->next: %p\n", this, curr, curr->next);
-	    curr = curr->next;
-	 }
-      }
-      return &curr->var[curr->pos++];
-   }
-
-   DLLLOCAL void uninstantiate(ExceptionSink* xsink) {
-      uninstantiateIntern();
-      curr->var[curr->pos].uninstantiate(xsink);
-   }
-
-   DLLLOCAL void uninstantiateSelf() {
-      uninstantiateIntern();
-      curr->var[curr->pos].uninstantiateSelf();
-   }
-
-   DLLLOCAL void uninstantiateIntern() {
-      if (!curr->pos) {
-	 if (curr->next) {
-	    //printf("this %p: del curr: %p, curr->next: %p\n", this, curr, curr->next);
-	    delete curr->next;
-	    curr->next = 0;
-	 }
-	 curr = curr->prev;
-         assert(curr);
-      }
-      --curr->pos;
-   }
-
-   DLLLOCAL LocalVarValue* find(const char* id) {
-      Block* w = curr;
-      while (true) {
-	 int p = w->pos;
-	 while (p) {
-	    if (w->var[--p].id == id && !w->var[p].skip && !w->var[p].frame_boundary)
-	       return &w->var[p];
-	 }
-	 w = w->prev;
-#ifdef DEBUG
-	 if (!w) {
-            printd(0, "ThreadLocalVariableData::find() this: %p no local variable '%s' (%p) on stack (pgm: %p) p: %d\n", this, id, id, getProgram(), p);
-            p = curr->pos - 1;
-            while (p >= 0) {
-               printd(0, "var p: %d: %s (%p) (skip: %d frame_boundary: %d)\n", p, curr->var[p].id, curr->var[p].id, curr->var[p].skip, curr->var[p].frame_boundary);
-               --p;
-            }
-         }
-#endif
-	 assert(w);
-      }
-      // to avoid a warning on most compilers - note that this generates a warning on recent versions of aCC!
-      return 0;
-   }
-
-   DLLLOCAL void pushFrameBoundary() {
-      //printd(5, "ThreadLocalVariableData::pushFrameBoundary()\n");
-      LocalVarValue* v = instantiate();
-      v->setFrameBoundary();
-   }
-
-   DLLLOCAL void popFrameBoundary() {
-      //printd(5, "ThreadLocalVariableData::popFrameBoundary()\n");
-      uninstantiateIntern();
-      assert(curr->var[curr->pos].frame_boundary);
-      curr->var[curr->pos].frame_boundary = false;
-   }
-
-   DLLLOCAL void getLocalVars(QoreHashNode& h, int frame, ExceptionSink* xsink) {
-      assert(frame >= 0);
-
-      // find requested frame
-      int cframe = 0;
-
-      bool in_frame = !frame;
-
-      Block* w = curr;
-      while (true) {
-	 int p = w->pos;
-	 while (p) {
-            --p;
-            const LocalVarValue& var = w->var[p];
-            if (var.frame_boundary) {
-               if (in_frame)
-                  break;
-               ++cframe;
-               if (frame == cframe)
-                  in_frame = true;
-               continue;
-            }
-
-            if (in_frame && !var.skip) {
-               ReferenceHolder<QoreHashNode> v(new QoreHashNode, xsink);
-               v->setKeyValue("type", new QoreStringNode("local"), xsink);
-               v->setKeyValue("value", var.evalValue(xsink).takeNode(), xsink);
-               h.setKeyValue(var.id, v.release(), xsink);
-            }
-	 }
-	 w = w->prev;
-         if (!w)
-            break;
-      }
-   }
-};
-
-class ThreadClosureVariableStack : public ThreadLocalData<ClosureVarValue*> {
-private:
-   DLLLOCAL void instantiateIntern(ClosureVarValue* cvar) {
-      //printd(5, "ThreadClosureVariableStack::instantiateIntern(%p = '%s') this: %p pgm: %p\n", cvar->id, cvar->id, this, getProgram());
-
-      if (curr->pos == QORE_THREAD_STACK_BLOCK) {
-	 if (curr->next)
-	    curr = curr->next;
-	 else {
-	    curr->next = new Block(curr);
-	    //printf("this: %p: add curr: %p, curr->next: %p\n", this, curr, curr->next);
-	    curr = curr->next;
-	 }
-      }
-      curr->var[curr->pos++] = cvar;
-   }
-
-public:
-   // marks all variables as finalized on the stack
-   DLLLOCAL void finalize(arg_vec_t*& cl) {
-      ThreadClosureVariableStack::iterator i(curr);
-      while (i.next()) {
-         AbstractQoreNode* n = i.get()->finalize();
-         if (n && n->isReferenceCounted()) {
-            if (!cl)
-               cl = new arg_vec_t;
-            cl->push_back(n);
-         }
-      }
-   }
-
-   // deletes everything on the stack
-   DLLLOCAL void del(ExceptionSink* xsink) {
-      while (curr->prev || curr->pos)
-         uninstantiate(xsink);
-   }
-
-   DLLLOCAL ClosureVarValue* instantiate(const char* id, const QoreTypeInfo* typeInfo, QoreValue& nval) {
-      ClosureVarValue* cvar = new ClosureVarValue(id, typeInfo, nval);
-      instantiateIntern(cvar);
-      return cvar;
-   }
-
-   DLLLOCAL void instantiate(ClosureVarValue* cvar) {
-      instantiateIntern(cvar);
-   }
-
-   DLLLOCAL void uninstantiateIntern() {
-#if 0
-      if (!curr->pos)
-         printd(5, "ThreadClosureVariableStack::uninstantiate() this: %p pos: %d %p %s\n", this, curr->prev->pos - 1, curr->prev->var[curr->prev->pos - 1]->id, curr->prev->var[curr->prev->pos - 1]->id);
-      else
-         printd(5, "ThreadClosureVariableStack::uninstantiate() this: %p pos: %d %p %s\n", this, curr->pos - 1, curr->var[curr->pos - 1]->id, curr->var[curr->pos - 1]->id);
-#endif
-      if (!curr->pos) {
-	 if (curr->next) {
-	    //printf("this %p: del curr: %p, curr->next: %p\n", this, curr, curr->next);
-	    delete curr->next;
-	    curr->next = 0;
-	 }
-	 curr = curr->prev;
-         assert(curr);
-      }
-      --curr->pos;
-   }
-
-   DLLLOCAL void uninstantiate(ExceptionSink* xsink) {
-      uninstantiateIntern();
-      assert(curr->var[curr->pos]);
-      curr->var[curr->pos]->deref(xsink);
-   }
-
-   DLLLOCAL ClosureVarValue* find(const char* id) {
-      printd(5, "ThreadClosureVariableStack::find() this: %p id: %p\n", this, id);
-      Block* w = curr;
-      while (true) {
-         int p = w->pos;
-         while (p) {
-            --p;
-            ClosureVarValue* rv = w->var[p];
-	    printd(5, "ThreadClosureVariableStack::find(%p '%s') this: %p checking %p '%s' skip: %d\n", id, id, this, rv ? rv->id : nullptr, rv ? rv->id : "n/a", rv ? rv->skip : false);
-	    if (rv && rv->id == id && !rv->skip) {
-	       printd(5, "ThreadClosureVariableStack::find(%p '%s') this: %p returning: %p\n", id, id, this, rv);
-	       return rv;
-	    }
-	 }
-	 w = w->prev;
-#ifdef DEBUG
-	 if (!w) {
-	    printd(0, "ThreadClosureVariableStack::find() this: %p no closure-bound local variable '%s' (%p) on stack (pgm: %p) p: %d curr->prev: %p\n", this, id, id, getProgram(), p, curr->prev);
-            p = curr->pos - 1;
-            while (p >= 0) {
-               ClosureVarValue* cvv = w->var[p];
-               printd(0, "var p: %d: %s (%p) (skip: %d)\n", p, cvv ? cvv->id : "frame boundary", cvv ? cvv->id : nullptr, cvv ? cvv->skip : false);
-               --p;
-            }
-         }
-#endif
-	 assert(w);
-      }
-      // to avoid a warning on most compilers - note that this generates a warning on aCC!
-      return 0;
-   }
-
-   DLLLOCAL cvv_vec_t* getAll() const {
-      cvv_vec_t* cv = 0;
-      Block* w = curr;
-      while (w) {
-         int p = w->pos;
-         while (p) {
-            --p;
-            ClosureVarValue* cvv = w->var[p];
-            // skip frame boundaries
-            if (!cvv)
-               continue;
-            if (!cv)
-               cv = new cvv_vec_t;
-            cv->push_back(cvv->refSelf());
-	 }
-	 w = w->prev;
-      }
-      //printd(5, "ThreadClosureVariableStack::getAll() this: %p cv: %p size: %d\n", this, cv, cv ? cv->size() : 0);
-      return cv;
-   }
-
-   DLLLOCAL void pushFrameBoundary() {
-      //printd(5, "ThreadClosureVariableStack::pushFrameBoundary()\n");
-      instantiateIntern(nullptr);
-   }
-
-   DLLLOCAL void popFrameBoundary() {
-      //printd(5, "ThreadClosureVariableStack::popFrameBoundary()\n");
-      uninstantiateIntern();
-      assert(!curr->var[curr->pos]);
-   }
-
-   DLLLOCAL void getLocalVars(QoreHashNode& h, int frame, ExceptionSink* xsink) {
-      assert(frame >= 0);
-
-      // find requested frame
-      int cframe = 0;
-
-      bool in_frame = !frame;
-
-      Block* w = curr;
-      while (true) {
-	 int p = w->pos;
-	 while (p) {
-            --p;
-            const ClosureVarValue* var = w->var[p];
-            if (!var) {
-               if (in_frame)
-                  break;
-               ++cframe;
-               if (frame == cframe)
-                  in_frame = true;
-               continue;
-            }
-
-            if (in_frame && !var->skip) {
-               ReferenceHolder<QoreHashNode> v(new QoreHashNode, xsink);
-               v->setKeyValue("type", new QoreStringNode("closure"), xsink);
-               v->setKeyValue("value", var->evalValue(xsink).takeNode(), xsink);
-               h.setKeyValue(var->id, v.release(), xsink);
-            }
-	 }
-	 w = w->prev;
-         if (!w)
-            break;
-      }
-   }
-};
+#include "qore/intern/ThreadlocalVariableData.h"
+#include "qore/intern/ThreadClosureVariableStack.h"
 
 struct ThreadLocalProgramData {
 private:
