+++ lib/QC_Program.qpp
@@ -1555,26 +1555,6 @@ bool Program::setThreadInit(*code init) [dom=THREAD_CONTROL] {
    return qore_program_private::setThreadInit(*p, init, xsink);
 }
 
-//! returns a hash of local variables for the given stack frame
-/** @par Example:
-    @code{.py}
-hash h = pgm.getLocalVars(int i);
-    @endcode
-
-    @param frame the stack frame starting from 0 (the current frame)
-
-    @return a hash of local variables for the given stack frame; if the stack frame is invalid, an empty hash is returned; hash keys are variable names; hash values are hashes with the following keys:
-    - \c "type": can be either "local" or "closure" for closure-bound local variables
-    - \c "value": the value of the variable
-
-    @see @ref Qore::get_local_vars()
-
-    @since %Qore 0.8.13
- */
-hash Program::getLocalVars(int frame) [dom=DEBUG_INSECURE;flags=RET_VALUE_ONLY] {
-   return thread_get_local_vars(frame, xsink);
-}
-
 //! returns a hash of global variables
 /** @par Example:
     @code{.py}
@@ -1589,24 +1569,6 @@ hash h = pgm.getGlobalVars();
  */
 hash Program::getGlobalVars() [flags=CONSTANT] {
    return p->getGlobalVars();
-}
-
-//! set the value of a local variable in the current stack frame
-/** @par Example:
-    @code
-pgm.setLocalVar("a", 1);
-    @endcode
-
-    @param name the name of the variable
-    @param value the value to assign
-
-    @throw INVALID-VARIABLE the variable is not present in the current stack frame
-
-    @note other exceptions could be thrown if the value cannot be assigned to the given variable
-
-    @since %Qore 0.8.13
- */
-nothing Program::setLocalVarValue(string name, any value) {
 }
 
 //! set the value of a global variable
