+++ lib/thread.cpp
@@ -890,6 +890,14 @@ QoreHashNode* thread_get_local_vars(int frame, ExceptionSink* xsink) {
    return rv.release();
 }
 
+int thread_set_local_var_value(const char* name, const QoreValue& val, ExceptionSink* xsink) {
+   return thread_data.get()->tlpd->lvstack.setVarValue(name, val, xsink);
+}
+
+int thread_set_closure_var_value(const char* name, const QoreValue& val, ExceptionSink* xsink) {
+   return thread_data.get()->tlpd->cvstack.setVarValue(name, val, xsink);
+}
+
 void parse_push_name(const char* name) {
    ThreadData* td = thread_data.get();
    td->pushName(name);
